+++ assets/src/edit-story/components/canvas/useUploadWithPreview.js
@@ -22,9 +22,10 @@ import { useCallback } from 'react';
 /**
  * Internal dependencies
  */
-import { useStory } from '../../app/story';
+import useStory from '../../app/story/useStory';
 import { useLocalMedia } from '../../app/media';
-import { DANGER_ZONE_HEIGHT, PAGE_HEIGHT, PAGE_WIDTH } from '../../constants';
+import { DANGER_ZONE_HEIGHT } from '../../constants';
+import useUpdateElementDimensions from '../../app/media/utils/useUpdateElementDimensions';
 import useInsertElement from './useInsertElement';
 
 function useUploadWithPreview() {
@@ -33,12 +34,10 @@ function useUploadWithPreview() {
     uploadVideoPoster: state.actions.uploadVideoPoster,
   }));
   const insertElement = useInsertElement();
-  const { updateElementsByResourceId, deleteElementsByResourceId } = useStory(
-    (state) => ({
-      updateElementsByResourceId: state.actions.updateElementsByResourceId,
-      deleteElementsByResourceId: state.actions.deleteElementsByResourceId,
-    })
-  );
+  const { updateElementDimensions } = useUpdateElementDimensions();
+  const { deleteElementsByResourceId } = useStory((state) => ({
+    deleteElementsByResourceId: state.actions.deleteElementsByResourceId,
+  }));
 
   const onUploadStart = useCallback(
     ({ resource }) => {
@@ -51,52 +50,22 @@ function useUploadWithPreview() {
     [insertElement]
   );
 
-  const updateElement = useCallback(
-    ({ id, resource }) => {
-      updateElementsByResourceId({
-        id,
-        properties: (el) => {
-          const hasChangedDimensions =
-            el.resource.width !== resource.width ||
-            el.resource.height !== resource.height;
-
-          if (!hasChangedDimensions) {
-            return {
-              type: resource.type,
-              resource,
-            };
-          }
-
-          return {
-            resource,
-            type: resource.type,
-            width: resource.width,
-            height: resource.height,
-            x: PAGE_WIDTH / 2 - resource.width / 2,
-            y: PAGE_HEIGHT / 2 - resource.height / 2,
-          };
-        },
-      });
-    },
-    [updateElementsByResourceId]
-  );
-
   const onUploadProgress = useCallback(
     ({ id, resource }) => {
-      updateElement({ id, resource });
+      updateElementDimensions({ id, resource });
     },
-    [updateElement]
+    [updateElementDimensions]
   );
 
   const onUploadSuccess = useCallback(
     ({ id, resource }) => {
-      updateElement({ id, resource });
+      updateElementDimensions({ id, resource });
 
       if (resource.type === 'video' && !resource.local && !resource.posterId) {
         uploadVideoPoster(resource.id, resource.src);
       }
     },
-    [updateElement, uploadVideoPoster]
+    [updateElementDimensions, uploadVideoPoster]
   );
 
   const onUploadError = useCallback(
